Control Statements
A number can be said as a strong number when the sum of the factorial of the individual digits is equal to the number.

For example, 145 is a strong number. 1! + 4! + 5! = 145. Write a program to check whether a given number is a strong number or not.

Input Format

Input consists of 1 integer.

Constraints

NA

Output Format

If it is a strong number print "Yes" or print "No".

Sample Input 0

145
Sample Output 0

Yes
Explanation 0

= 1! + 4! +5! = 1+24+120 = 145

Sample Input 1

123
Sample Output 1

No
Explanation 1

= 1! + 2! + 3! = 1+2+6 = 9