Arrays 1D
Mahe and Mani are playing a puzzle game with a given set of numbers. They need to find the sum of the odd number and the even number. Write a program to help them to solve the puzzle game.

Input Format

Input consists of n+1 integers. The first integer corresponds to ‘n’, the size of the array. The next ‘n’ integers correspond to the elements in the array. Assume that the maximum value of n is 15.

Constraints

NA

Output Format

Refer to the sample output for details.

Sample Input 0

5
2
3
6
8
-1
Sample Output 0

The sum of the even numbers in the array is 16
The sum of the odd numbers in the array is 2
Sample Input 1

4
1
2
3
-1
Sample Output 1

The sum of the even numbers in the array is 2
The sum of the odd numbers in the array is 3